practices_3+/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
Jinja2==3.1.6
python-dotenv==1.2.2
mysql-connector-python